chore: update alloy deps

+++ crates/preconfer/src/pricer.rs
@@ -11,7 +11,7 @@ use crate::error::PricerError;
 pub trait PreconfPricer {
     fn get_optimal_base_gas_fee(
         &self,
-    ) -> impl std::future::Future<Output = eyre::Result<u128, PricerError>> + Send;
+    ) -> impl std::future::Future<Output = eyre::Result<u64, PricerError>> + Send;
     /// Simply scale up the current base fee by 10% per block
     fn price_preconf(
         &self,
@@ -21,7 +21,7 @@ pub trait PreconfPricer {
         Self: std::marker::Sync,
     {
         async move {
-            let current_base_fee: u128 = self.get_optimal_base_gas_fee().await?;
+            let current_base_fee: u64 = self.get_optimal_base_gas_fee().await?;
             let current_base_fee_f64: f64 = current_base_fee as f64;
             let block_lookahead_f64: f64 = block_lookahead as f64;
             let projected_base_fee_f64: f64 =
@@ -43,12 +43,12 @@ impl TaiyiFeePricer {
 }
 
 impl PreconfPricer for TaiyiFeePricer {
-    async fn get_optimal_base_gas_fee(&self) -> eyre::Result<u128, PricerError> {
+    async fn get_optimal_base_gas_fee(&self) -> eyre::Result<u64, PricerError> {
         let response = reqwest::get(self.url.clone()).await?;
         let body = response.bytes().await?;
 
         let body_str = String::from_utf8_lossy(&body);
-        let res = body_str.parse::<u128>().map_err(|e| PricerError::ParseError(e.to_string()))?;
+        let res = body_str.parse::<u64>().map_err(|e| PricerError::ParseError(e.to_string()))?;
         Ok(res)
     }
 }
@@ -78,7 +78,7 @@ where
     T: Transport + Clone,
     P: Provider<T, Ethereum> + Clone,
 {
-    async fn get_optimal_base_gas_fee(&self) -> eyre::Result<u128, PricerError> {
+    async fn get_optimal_base_gas_fee(&self) -> eyre::Result<u64, PricerError> {
         let block = self
             .provider
             .get_block(BlockId::Number(BlockNumberOrTag::Latest), BlockTransactionsKind::Hashes)
